tasks/lookingsus/app/requirements.txt
aiohttp
aiohttp_jinja2
PyJWT
